pyproject.toml
[tool.poetry]
name = "friendly-gql"
version = "0.1.0"
description = "A GraphQL server implementation using Python annotations"
authors = ["Adam Brenecki"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = ">=3.6"
graphql-core-next = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^3.9"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
